requirements.txt
pandas
numpy
python-dotenv
ccxt
ta
scikit-learn
apscheduler
pyupbit
streamlit
matplotlib
openpyxl
requests